requirements.txt
accelerate
torch
pillow
transformers
notebook 
tensorflow 
opencv-python